+++ packages/gatsby-cli/src/init-starter.js
@@ -206,8 +206,10 @@ const getPaths = async (starterPath: string, rootPath: string) => {
       },
     ])
     // exit gracefully if responses aren't provided
-    if (!response.starter || !response.path) {
-      process.exit()
+    if (!response.starter || !response.path.trim()) {
+      throw new Error(
+        `Please mention both starter package and project name along with path(if its not in the root)`
+      )
     }
 
     selectedOtherStarter = response.starter === `different`
@@ -224,6 +226,14 @@ const getPaths = async (starterPath: string, rootPath: string) => {
 
 type InitOptions = {
   rootPath?: string,
+}
+
+const successMessage = path => {
+  report.info(`
+Your new Gatsby site has been successfully bootstrapped. Start developing it by running:
+  $ cd ${path}
+  $ gatsby develop
+`)
 }
 
 /**
@@ -244,27 +254,49 @@ module.exports = async (starter: string, options: InitOptions = {}) => {
     opn(`https://gatsby.dev/starters?v=2`)
     return
   }
-
   if (urlObject.protocol && urlObject.host) {
     trackError(`NEW_PROJECT_NAME_MISSING`)
-    report.panic(
-      `It looks like you forgot to add a name for your new project. Try running instead "gatsby new new-gatsby-project ${rootPath}"`
-    )
+
+    const isStarterAUrl =
+      starter && !url.parse(starter).hostname && !url.parse(starter).protocol
+
+    if (/gatsby-starter/gi.test(rootPath) && isStarterAUrl) {
+      report.panic({
+        id: `11610`,
+        context: {
+          starter,
+          rootPath,
+        },
+      })
+      return
+    }
+    report.panic({
+      id: `11611`,
+      context: {
+        rootPath,
+      },
+    })
     return
   }
 
   if (!isValid(rootPath)) {
-    report.panic(
-      `Could not create a project in "${sysPath.resolve(
-        rootPath
-      )}" because it's not a valid path`
-    )
+    report.panic({
+      id: `11612`,
+      context: {
+        path: sysPath.resolve(rootPath),
+      },
+    })
     return
   }
 
   if (existsSync(sysPath.join(rootPath, `package.json`))) {
     trackError(`NEW_PROJECT_IS_NPM_PROJECT`)
-    report.panic(`Directory ${rootPath} is already an npm project`)
+    report.panic({
+      id: `11613`,
+      context: {
+        rootPath,
+      },
+    })
     return
   }
 
@@ -275,5 +307,6 @@ module.exports = async (starter: string, options: InitOptions = {}) => {
   })
   if (hostedInfo) await clone(hostedInfo, rootPath)
   else await copy(starterPath, rootPath)
+  successMessage(rootPath)
   trackCli(`NEW_PROJECT_END`)
 }