+++ packages/gatsby-plugin-offline/src/gatsby-ssr.js
@@ -1,3 +1,5 @@
+import React from "react"
+
 export const onPreRenderHTML = ({
   getHeadComponents,
   pathname,
@@ -19,4 +21,21 @@ export const onPreRenderHTML = ({
   )
 
   replaceHeadComponents(filteredHeadComponents)
+}
+
+export const onRenderBody = ({ pathname, setHeadComponents }) => {
+  if (pathname !== `/offline-plugin-app-shell-fallback/`) {
+    return
+  }
+
+  setHeadComponents([
+    <noscript key="disable-offline-shell">
+      <link
+        rel="preload"
+        as="fetch"
+        href="/.gatsby-plugin-offline:disableOfflineShell"
+      />
+      <meta httpEquiv="refresh" content="0;url=" />
+    </noscript>,
+  ])
 }